Made EActivities and model classes available in the root package
