+++ pyeactivities/__init__.py
@@ -1,1 +1,15 @@
 __version__ = "0.1.1"
+
+from pyeactivities.eactivities import (
+    EActivities,
+    CSP,
+    Member,
+    CommitteeMember,
+    OnlineSale,
+    Customer,
+    VAT,
+    Product,
+    ProductLine,
+    Account,
+    Activity,
+)
